index.js
I am attempting to create a second database to handle the offline tasks.  The first database is giving me errors and I am pretty sure it is not set up optimally.  I am going to set up the second database, attempt to use it to accomplish the same tasks as the first, and see if I come up against the same errors and see if I learn anything